// src/app/share/swagger-auto-gen/model/dtosServiceInstanceDetailsDto.ts
/**
 * Operator Automation Backend API
 * Operator Automation Backend API overview.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DtosServiceInstanceActionGroupDto } from './dtosServiceInstanceActionGroupDto';


export interface DtosServiceInstanceDetailsDto { 
    actionGroups?: Array<DtosServiceInstanceActionGroupDto>;
    name?: string;
    status?: string;
    type?: string;
}